Add multi role support to SetAufstellung
The endpoint api/SetAfstelllung was not supporting the "new" multi
role feature, so it required an update.

Changed from conversion to number to string in the api/contoller.
Updated the parameter roleId of element/setCompleteElement by
changing the type to string and renaming it to roles as it better
fits the new purpose.

The parameter name of the external api endpoint was not changed to
preserve backwards comptibility.