Add the parameter "reportLink" to api/aufstellungen
The new parameter "reportLink" is used to provide the dps report link
for the encounter via the api/aufstellungen endpoint. This allows to
set also the report in one request with no long hold up. Also this
speeds up the timer required for already uploaded reports, since no
second upload needs to be performed.